# openapi/oh.yaml
openapi: 3.1.0
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
- url: http://API_HOST:API_PORT
security:
- bearerAuth: []
...
paths:
...
  /medicaltypes:
    get:
      tags:
      - Medical Types
      operationId: getMedicalTypes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalTypeDTO"
      security:
      - bearerAuth: []
    put:
      tags:
      - Medical Types
      operationId: updateMedicalType
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalTypeDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalTypeDTO"
      security:
      - bearerAuth: []
    post:
      tags:
      - Medical Types
      operationId: createMedicalType
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalTypeDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalTypeDTO"
      security:
      - bearerAuth: []
...
  /examtypes/{code}:
    put:
      tags:
      - Exam Types
      operationId: updateExamType
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamTypeDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamTypeDTO"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Exam Types
      operationId: deleteExamType
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
...
  /diseasetypes:
    get:
      tags:
      - Disease Types
      operationId: getAllDiseaseTypes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiseaseTypeDTO"
      security:
      - bearerAuth: []
    put:
      tags:
      - Disease Types
      operationId: updateDiseaseType
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiseaseTypeDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiseaseTypeDTO"
      security:
      - bearerAuth: []
    post:
      tags:
      - Disease Types
      operationId: newDiseaseType
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiseaseTypeDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiseaseTypeDTO"
      security:
      - bearerAuth: []
...
  /examtypes:
    get:
      tags:
      - Exam Types
      operationId: getExamTypes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamTypeDTO"
      security:
      - bearerAuth: []
    post:
      tags:
      - Exam Types
      operationId: newExamType
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamTypeDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamTypeDTO"
      security:
      - bearerAuth: []
...
  /medstockmovementtypes/check/{code}:
    get:
      tags:
      - Medical Stock Movement Type
      operationId: isCodeUsed
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
  /medicaltypes/check/{code}:
    get:
      tags:
      - Medical Types
      operationId: isCodeUsed_1
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
...
  /medicaltypes/{code}:
    delete:
      tags:
      - Medical Types
      operationId: deleteMedicalType
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
  /diseasetypes/{code}:
    delete:
      tags:
      - Disease Types
      operationId: deleteDiseaseType
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    DiseaseTypeDTO:
      type: object
      description: Class representing a disease type
      properties:
        code:
          type: string
          description: Disease type code
          maxLength: 2
        description:
          type: string
          description: Disease type description
          maxLength: 110
      required:
      - code
      - description
...
    MedicalTypeDTO:
      type: object
      properties:
        code:
          type: string
          description: Code of the medical type
          example: M
          maxLength: 1
        description:
          type: string
          description: Description of the medical type
          example: Medical material
          maxLength: 30
      required:
      - code
      - description
...
    ExamTypeDTO:
      type: object
      properties:
        code:
          type: string
          description: Exam Type Code
          example: HB
        description:
          type: string
          description: Exam Type Description
          example: 1.Haematology
          maxLength: 50
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT